refactor(dto): insert request/response folders
