+++ backend/src/main/java/com/example/worksync/dto/requests/AuthDTO.java
@@ -1,4 +1,4 @@
-package com.example.worksync.dto;
+package com.example.worksync.dto.requests;
 
 public class AuthDTO {
