+++ backend/src/main/java/com/example/worksync/dto/responses/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.example.worksync.dto;
+package com.example.worksync.dto.responses;
 
 public class LoginResponseDTO {
